lib/EngineDesign/requirements.txt
numpy
scipy
pandas
matplotlib
pydantic
PyYAML
rocketcea
rocketpy
streamlit
plotly
ezdxf
cma
CoolProp
# FastAPI backend
fastapi
uvicorn[standard]
python-multipart
